chore(lint): rebuild the quality-debt baseline from measured findings
Landing four paydown waves through a line-based merge corrupted several
multi-line baseline entries, because a rule line like `"file-too-long",` is
not unique in the file. Regenerated the block from a real `poly lint` run
with the baseline empty: 291 files, 626 findings, down from 422/865.

Also fixes the GLiNER candle pipeline, whose extracted helper declared
`&[usize]` for a field that is `Vec<i64>` so the candle feature did not
compile, and allowlists the German and Dutch words the new PDF reproducer
fixtures introduced.

+++ crates/xberg-gliner/src/candle/pipeline.rs
@@ -108,7 +108,7 @@ fn gather_and_span_rep(
 /// [`run_pipeline`] to keep that function under the workspace line-count limit. ~keep
 fn compute_pred_count_and_struct_proj(
     hidden: &Tensor,
-    schema_positions: &[usize],
+    schema_positions: &[i64],
     heads: &crate::candle::heads::AllHeads,
     device: &candle_core::Device,
 ) -> crate::candle::Result<(usize, Option<Tensor>)> {
